fix: skeleton refactor (#353)
