+++ src/Components/Home/Home.css
@@ -5,6 +5,3 @@
 .p-avatar:hover {
   transform: scale(1.2);
 }
-.skeleton{
-  display: none;
-}
